fix(web): reuse agent flavor icon in session header
